+++ example/web_frontend/src/gui/testgui/controller/none.js
@@ -94,7 +94,7 @@ module.exports = (...args) => {
           js.locally.push("/public/assets/js/model/sample.js");
           mjs.initialize["sample.init"] = [[layer.params.title]];
           less.engine.remotely = "/npm/less@4.2.0/dist/less.min.js";
-          less.style.locally = ["/public/assets/less/plugin/wi.less"];
+          less.style.locally = ["/less/plugin/wi.less"];
 
           return response;
         } catch (error) {