+++ Changelog.md
@@ -5,11 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.3] - 2024-07-26
+## [1.1.4] - 2024-08-26
 
 ### Summary
 
 - Work with framework OriCommJS 1.1.6 and above.
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.1.4]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.1.4
+
+## [1.1.3] - 2024-07-26
+
+### Summary
+
+- Work with framework OriCommJS 1.1.4 and above.
 
 ### Added